% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_this.R
\name{keep_this}
\alias{keep_this}
\title{Delete everything in the R environment BUT
keep the elements in the function.
input the elements as: "a","b",...}
\usage{
keep_this(elements)
}
\description{
Delete everything in the R environment BUT
keep the elements in the function.
input the elements as: "a","b",...
}
