% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_test.R
\name{mi_test}
\alias{mi_test}
\title{Delete everything in the R environment BUT
keep the elements in the function.
input the elements as: "a","b",...}
\usage{
mi_test(cosa)
}
\arguments{
\item{cosa}{single object or a list of them to be removed "a","b",...}
}
\description{
Delete everything in the R environment BUT
keep the elements in the function.
input the elements as: "a","b",...
}
