% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ari_year.R
\name{theme_ari_year}
\alias{theme_ari_year}
\title{Theme template for ggplot
Use for yearly figures showing trends}
\usage{
theme_ari_year(...)
}
\description{
Theme template for ggplot
Use for yearly figures showing trends
}
